+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/anthropic
 
+## 3.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider@3.0.6
+  - @zenning/provider-utils@4.0.8
+
+## 3.0.10
+
+### Patch Changes
+
+- Updated dependencies [35e1a30]
+  - @zenning/provider@3.0.5
+  - @zenning/provider-utils@4.0.7
+
 ## 3.0.9
 
 ### Patch Changes