+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/deepseek
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider@3.0.6
+  - @zenning/provider-utils@4.0.8
+
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [35e1a30]
+  - @zenning/provider@3.0.5
+  - @zenning/provider-utils@4.0.7
+
 ## 2.0.6
 
 ### Patch Changes