+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/google-vertex
 
+## 4.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @zenning/provider@3.0.6
+  - @zenning/anthropic@3.0.11
+  - @zenning/google@3.0.8
+  - @zenning/provider-utils@4.0.8
+
+## 4.0.10
+
+### Patch Changes
+
+- Updated dependencies [35e1a30]
+  - @zenning/provider@3.0.5
+  - @zenning/provider-utils@4.0.7
+  - @zenning/anthropic@3.0.10
+  - @zenning/google@3.0.7
+
 ## 4.0.9
 
 ### Patch Changes