+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/valibot
 
+## 2.0.8
+
+### Patch Changes
+
+- @zenning/provider-utils@4.0.8
+
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [35e1a30]
+  - @zenning/provider-utils@4.0.7
+
 ## 2.0.6
 
 ### Patch Changes